split qsvt.py
