+++ src/bequem/nodes/qsvt/amplitude_amplification.py
@@ -1,0 +1,14 @@
+import numpy as np
+
+from bequem.nodes import Node
+
+from bequem.nodes.qsvt.qsvt import QSVT
+
+
+class AmplitudeAmplificiation(QSVT):
+    def __init__(self, A: Node, iterations: int):
+        angles = None
+        raise NotImplementedError
+        super().__init__(A, angles)
+        self.A = A
+        self.iterations = iterations
